Added view subscriptions method

+++ Model/src/main/java/event/rec/service/requests/ViewFavouriteRequest.java
@@ -1,5 +1,5 @@
 package event.rec.service.requests;
 
 public record ViewFavouriteRequest(Long userId,
-                                   Long page,
-                                   Long size) {}
+                                   Integer page,
+                                   Integer size) {}
